seap_api/requirements.txt
django
djangorestframework
django-mongoengine
python-dotenv
mongoengine
werkzeug
PyJWT
django-cors-headers
psutil